fix dgl nas import
